pyproject.toml
[project]
name = "blender-agent-simple"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.6"
dependencies = [
    "google-generativeai>=0.8.6",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
]
